+++ Libraries/LibJS/Bytecode/Executable.cpp
@@ -5,6 +5,7 @@
  */
 
 #include <LibJS/Bytecode/BasicBlock.h>
+#include <LibJS/Runtime/Value.h>
 #include <LibJS/Bytecode/Executable.h>
 #include <LibJS/Bytecode/Instruction.h>
 #include <LibJS/Bytecode/RegexTable.h>
